adding team specific routes
