+++ routes/users.js
@@ -1,6 +1,6 @@
 var express = require("express");
 const generateAccessToken = require("../config/jwt_generator");
-const User = require("../models/User.model");
+const User = require("../models/team_d_models/User.model");
 var router = express.Router();
 const createError = require("http-errors");
 
@@ -22,47 +22,6 @@ router.get("/users/:id", async function (req, res) {
   res.json(user);
 });
 
-router.patch("/add-cart/:id", async (req, res, next) => {
-  const user = await User.findById(req.params.id);
-  const { product, quantity } = req.body;
-  user.cartItem.push({
-    product,
-    quantity,
-  });
-  const savedUser = await user.save();
-  const suser = await User.findById(req.params.id).populate({
-    path: "cartItem",
-    populate: "product",
-  });
-  res.send(suser);
-});
-
-router.patch("/handle-cart-quantity/:id", async (req, res) => {
-  const { cartId, quantity } = req.body;
-  const user = await User.findById(req.params.id);
-  const cart = user.cartItem.find((item) => item._id.toString() === cartId);
-  cart.quantity = quantity || 1;
-  await user.save();
-  const suser = await User.findById(req.params.id).populate({
-    path: "cartItem",
-    populate: "product",
-  });
-  res.send(suser);
-});
-
-router.patch("/remove-cart/:id", async (req, res, next) => {
-  const user = await User.findById(req.params.id);
-  const { _id } = req.body;
-  user.cartItem.pull({
-    _id,
-  });
-  await user.save();
-  const suser = await User.findById(req.params.id).populate({
-    path: "cartItem",
-    populate: "product",
-  });
-  res.send(suser);
-});
 
 router.post("/auth/register", async function (req, res, next) {
   try {
